feature：发布时间一天 / 一周内新闻按浏览量降序的API，支持分页

+++ news-service/src/main/kotlin/com/example/joyread/news/repository/converter/NewsContentConverter.kt
@@ -25,7 +25,7 @@ class NewsContentConverter : Converter<Document, ParagraphBlock> {
                 val text = source.get("text") as String
                 val levelSource = source.get("level") as Int
 
-                val level = levelSource.toUShort()
+                val level = levelSource.toUByte()
                 HeadingBlock(text, level)
             }
